requirements.txt
edge-tts
playsound==1.2.2
sounddevice
SpeechRecognition
numpy
vosk        # optional (only if you’ll use offline STT later)

python-dotenv
twilio
